[Fix] Main Category onClick Event
- 메인 카테고리 클릭 이벤트
